project complete

+++ WishList/Controllers/ItemController.cs
@@ -1,30 +1,42 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
+
+
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
+
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
-
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            var model = _context.Items.Where(e => e.User.Id == user.Id).ToList();
             return View("Index", model);
         }
+
 
         [HttpGet]
         public IActionResult Create()
         {
             return View("Create");
         }
+
 
         [HttpPost]
         public IActionResult Create(Models.Item item)
@@ -34,9 +46,14 @@ namespace WishList.Controllers
             return RedirectToAction("Index");
         }
 
+
+
         public IActionResult Delete(int id)
         {
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if (item.User.Id != user.Id)
+                return Unauthorized();
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
